feat: add flag for reading specified puzzle part
Add a `--part` flag in the `read` subcommand to display/read the
specified part of the puzzle.

When part 1 is included in display (either by directly specifying
`--part 1`, or by not specifying `--part` flag which defaults to both
part 1 & 2), the `--- Part One ---` subheader is added in the output.
This makes display consistent with the output for part 2
(`--- Part Two ---`, which is included by default), and adds better
indicator for the puzzle sections.

Fixes: #27